# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/v1/post": {
            "post": {
                "tags": [
                    "게시판 글"
                ],
                "summary": "게시판 글 작성",
                "description": "게시판 글 작성",
                "operationId": "postCreate",
                "requestBody": {
                    "description": "",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/postCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successfully Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "successfully Modified"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "failed registered",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "properties": {
                                                "statusCode": {
                                                    "properties": {
                                                        "20010": {
                                                            "description": "이미 존재하는 type 입니다.",
                                                            "properties": {
                                                                "key": {
                                                                    "description": "type",
                                                                    "type": "string",
                                                                    "example": "type"
                                                                },
                                                                "message": {
                                                                    "type": "string"
                                                                }
                                                            },
                                                            "type": "object"
                                                        },
                                                        "20401": {
                                                            "description": "옵션 및 옵션 값을 확인해주세요.",
                                                            "properties": {
                                                                "key": {
                                                                    "description": "options.editor.value",
                                                                    "type": "string",
                                                                    "example": "options.editor.value"
                                                                },
                                                                "message": {
                                                                    "type": "string"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "postCreate": {
                "required": [
                    "boardNo",
                    "userNo",
                    "title",
                    "content"
                ],
                "properties": {
                    "boardNo": {
                        "description": "게시판 고유 번호",
                        "type": "string",
                        "example": 1
                    },
                    "title": {
                        "description": "게시글 제목",
                        "type": "string",
                        "example": "게시글 입니다."
                    },
                    "content": {
                        "description": "게시글 내용",
                        "type": "string",
                        "example": "게시글 내용입니다."
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "davinci_auth": {
                "type": "http",
                "scheme": "bearer"
            },
            "sample_auth": {
                "type": "oauth2",
                "flows": {
                    "implicit": {
                        "authorizationUrl": "http://petstore.swagger.io/oauth/dialog",
                        "scopes": {
                            "read:pets": "read your pets",
                            "write:pets": "modify pets in your account"
                        }
                    }
                }
            }
        }
    },
...
    "security": [
        []
    ]
}